add some magic
